requirements.txt
requests
python-dotenv
joblib
scikit-learn
pandas
openai
websockets
fastapi
uvicorn
pytest
